refactor(admin): auth pages to TS
