+++ packages/core/admin/admin/src/hooks/useEnterprise.ts
@@ -12,21 +12,26 @@ function isEnterprise() {
   return window.strapi.isEE;
 }
 
-export interface UseEnterpriseOptions<TCEData, TEEData, TCombinedData> {
+export interface UseEnterpriseOptions<
+  TCEData = unknown,
+  TEEData = unknown,
+  TCombinedData = TEEData
+> {
   defaultValue?: TCEData | TEEData | null;
-  combine?: (ceData: TCEData, eeData: TEEData) => TCEData | TEEData | TCombinedData;
+  combine?: (ceData: TCEData, eeData: TEEData) => TCombinedData;
   enabled?: boolean;
 }
 
-export function useEnterprise<TCEData, TEEData, TCombinedData = unknown>(
+export function useEnterprise<TCEData = unknown, TEEData = unknown, TCombinedData = TEEData>(
   ceData: TCEData,
   eeCallback: () => Promise<TEEData>,
   {
     defaultValue = null,
-    combine = <C, E>(ceData: C, eeData: E) => eeData,
+    // @ts-expect-error – TODO: fix this type
+    combine = (ceData, eeData) => eeData,
     enabled = true,
   }: UseEnterpriseOptions<TCEData, TEEData, TCombinedData> = {}
-) {
+): null | TCEData | TEEData | TCombinedData {
   const eeCallbackRef = useCallbackRef(eeCallback);
   const combineCallbackRef = useCallbackRef(combine);
