+++ packages/core/admin/admin/src/pages/Auth/components/FieldActionWrapper.tsx
@@ -11,4 +11,4 @@ const FieldActionWrapper = styled(FieldAction)`
   }
 `;
 
-export default FieldActionWrapper;
+export { FieldActionWrapper };
