+++ packages/core/admin/admin/src/pages/Auth/components/ResetPassword.tsx
@@ -1,73 +1,110 @@
-import React, { useState } from 'react';
+import * as React from 'react';
 
 import { Box, Button, Flex, Main, TextInput, Typography } from '@strapi/design-system';
-import { Form, Link } from '@strapi/helper-plugin';
+import { Link } from '@strapi/design-system/v2';
+import { Form, auth, translatedErrors, useFetchClient, useQuery } from '@strapi/helper-plugin';
 import { Eye, EyeStriked } from '@strapi/icons';
 import { Formik } from 'formik';
-import PropTypes from 'prop-types';
 import { useIntl } from 'react-intl';
+import { useMutation } from 'react-query';
+import { NavLink, useHistory } from 'react-router-dom';
 import styled from 'styled-components';
+import * as yup from 'yup';
 
-import { Logo } from '../../../../components/UnauthenticatedLogo';
+import { ResetPassword } from '../../../../../shared/contracts/authentication';
+import { Logo } from '../../../components/UnauthenticatedLogo';
 import {
   Column,
   LayoutContent,
   UnauthenticatedLayout,
-} from '../../../../layouts/UnauthenticatedLayout';
-import FieldActionWrapper from '../FieldActionWrapper';
+} from '../../../layouts/UnauthenticatedLayout';
 
-const PasswordInput = styled(TextInput)`
-  ::-ms-reveal {
-    display: none;
-  }
-`;
+import { FieldActionWrapper } from './FieldActionWrapper';
 
-const ForgotPassword = ({ onSubmit, schema }) => {
-  const [passwordShown, setPasswordShown] = useState(false);
-  const [confirmPasswordShown, setConfirmPasswordShown] = useState(false);
+const REGISTER_SCHEMA = yup.object().shape({
+  password: yup
+    .string()
+    .min(8, translatedErrors.minLength)
+    .matches(/[a-z]/, 'components.Input.error.contain.lowercase')
+    .matches(/[A-Z]/, 'components.Input.error.contain.uppercase')
+    .matches(/\d/, 'components.Input.error.contain.number')
+    .required(translatedErrors.required),
+  confirmPassword: yup
+    .string()
+    .oneOf([yup.ref('password'), null], 'components.Input.error.password.noMatch')
+    .required(translatedErrors.required),
+});
+
+const ResetPassword = () => {
+  const [passwordShown, setPasswordShown] = React.useState(false);
+  const [confirmPasswordShown, setConfirmPasswordShown] = React.useState(false);
   const { formatMessage } = useIntl();
+  const { post } = useFetchClient();
+  const { push } = useHistory();
+  const query = useQuery();
+
+  const { mutate, isError } = useMutation(
+    async (body: Pick<ResetPassword.Request['body'], 'password'>) => {
+      const {
+        data: { data },
+      } = await post<ResetPassword.Response>('/admin/reset-password', {
+        ...body,
+        resetPasswordToken: query.get('code'),
+      });
+
+      return data;
+    },
+    {
+      onSuccess(data) {
+        if (data) {
+          const { token, user } = data;
+          auth.setToken(token, false);
+          auth.setUserInfo(user, false);
+
+          // Redirect to the homePage
+          push('/');
+        }
+      },
+    }
+  );
 
   return (
     <UnauthenticatedLayout>
       <Main>
         <LayoutContent>
+          <Column>
+            <Logo />
+            <Box paddingTop={6} paddingBottom={7}>
+              <Typography as="h1" variant="alpha">
+                {formatMessage({
+                  id: 'global.reset-password',
+                  defaultMessage: 'Reset password',
+                })}
+              </Typography>
+            </Box>
+            {isError ? (
+              <Typography id="global-form-error" role="alert" tabIndex={-1} textColor="danger600">
+                {formatMessage({
+                  id: 'notification.error',
+                  defaultMessage: 'An error occurred',
+                })}
+              </Typography>
+            ) : null}
+          </Column>
           <Formik
             enableReinitialize
             initialValues={{
               password: '',
               confirmPassword: '',
             }}
-            onSubmit={onSubmit}
-            validationSchema={schema}
+            onSubmit={(values) => {
+              mutate(values);
+            }}
+            validationSchema={REGISTER_SCHEMA}
             validateOnChange={false}
           >
             {({ values, errors, handleChange }) => (
-              <Form noValidate>
-                <Column>
-                  <Logo />
-                  <Box paddingTop={6} paddingBottom={7}>
-                    <Typography as="h1" variant="alpha">
-                      {formatMessage({
-                        id: 'global.reset-password',
-                        defaultMessage: 'Reset password',
-                      })}
-                    </Typography>
-                  </Box>
-                  {errors.errorMessage && (
-                    <Typography
-                      id="global-form-error"
-                      role="alert"
-                      tabIndex={-1}
-                      textColor="danger600"
-                    >
-                      {formatMessage({
-                        id: errors.errorMessage,
-                        defaultMessage: 'An error occurred',
-                      })}
-                    </Typography>
-                  )}
-                </Column>
-
+              <Form>
                 <Flex direction="column" alignItems="stretch" gap={6}>
                   <PasswordInput
                     name="password"
@@ -154,7 +191,7 @@ const ForgotPassword = ({ onSubmit, schema }) => {
                     })}
                     type={confirmPasswordShown ? 'text' : 'password'}
                   />
-                  <Button fullwidth type="submit">
+                  <Button fullWidth type="submit">
                     {formatMessage({
                       id: 'global.change-password',
                       defaultMessage: 'Change password',
@@ -167,7 +204,8 @@ const ForgotPassword = ({ onSubmit, schema }) => {
         </LayoutContent>
         <Flex justifyContent="center">
           <Box paddingTop={4}>
-            <Link to="/auth/login">
+            {/* @ts-expect-error – error with inferring the props from the as component */}
+            <Link as={NavLink} to="/auth/login">
               {formatMessage({ id: 'Auth.link.ready', defaultMessage: 'Ready to sign in?' })}
             </Link>
           </Box>
@@ -177,15 +215,10 @@ const ForgotPassword = ({ onSubmit, schema }) => {
   );
 };
 
-ForgotPassword.defaultProps = {
-  onSubmit: (e) => e.preventDefault(),
-};
+const PasswordInput = styled(TextInput)`
+  ::-ms-reveal {
+    display: none;
+  }
+`;
 
-ForgotPassword.propTypes = {
-  onSubmit: PropTypes.func,
-  schema: PropTypes.shape({
-    type: PropTypes.string.isRequired,
-  }).isRequired,
-};
-
-export default ForgotPassword;
+export { ResetPassword };
