--- packages/core/admin/admin/src/pages/AuthPage/init.js
@@ -1,5 +1,0 @@
-const init = (initialState) => {
-  return initialState;
-};
-
-export default init;
