--- packages/core/admin/ee/admin/src/pages/AuthPage/constants.js
@@ -1,12 +1,0 @@
-import Providers from './components/Providers';
-
-export const FORMS = {
-  providers: {
-    Component: Providers,
-    endPoint: null,
-    fieldsToDisable: [],
-    fieldsToOmit: [],
-    schema: null,
-    inputsPrefix: '',
-  },
-};
