+++ packages/core/admin/ee/admin/src/pages/AuthPage/constants.ts
@@ -1,0 +1,11 @@
+import type { ComponentType } from 'react';
+
+import { Providers } from './components/Providers';
+
+type AuthType = 'providers';
+
+type FormDictionary = Record<AuthType, ComponentType>;
+
+export const FORMS = {
+  providers: Providers,
+} satisfies FormDictionary;
