+++ packages/core/admin/package.json
@@ -179,6 +179,7 @@
     "@testing-library/user-event": "14.4.3",
     "@types/find-root": "1.1.3",
     "@types/invariant": "2.2.36",
+    "@types/js-cookie": "3.0.5",
     "@types/pluralize": "0.0.32",
     "@types/webpack-bundle-analyzer": "4.6.2",
     "@types/webpack-hot-middleware": "2.25.8",
